fixed Makefile relink + ft_memset size_t count

+++ ft_memset.c
@@ -6,15 +6,15 @@
 /*   By: dkrecisz <dkrecisz@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/11/02 13:12:43 by dkrecisz      #+#    #+#                 */
-/*   Updated: 2019/11/21 00:24:14 by dkrecisz      ########   odam.nl         */
+/*   Updated: 2020/09/08 20:37:44 by dkrecisz      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "./include/libft.h"
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int		i;
+	size_t	i;
 
 	i = 0;
 	while (len)
